마이페이지 구독한 지역에서 열리는 행사 조회 api
