+++ src/main/java/kori/tour/member/adapter/in/MemberApiParser.java
@@ -4,18 +4,22 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import kori.tour.global.data.area_code.AreaCodeRegistry;
+import kori.tour.member.adapter.in.api.out.*;
+import kori.tour.tour.domain.Tour;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Component;
 
 import kori.tour.global.data.area_code.Area;
-import kori.tour.member.adapter.in.api.out.AreaDto;
-import kori.tour.member.adapter.in.api.out.MyPage;
-import kori.tour.member.adapter.in.api.out.SubAreaDto;
-import kori.tour.member.adapter.in.api.out.SubscriptionsResponse;
 import kori.tour.member.domain.Member;
 import kori.tour.member.domain.Subscription;
 
 @Component
+@RequiredArgsConstructor
 class MemberApiParser {
+
+    private final AreaCodeRegistry areaCodeRegistry;
 
     public MyPage mapToMyPageResponse(Member member) {
         return new MyPage(member.getPlatformInfo().getPlatformEmail(), member.getEmailSubscribe());
@@ -46,6 +50,25 @@ class MemberApiParser {
         // 3. 최종 응답
         return new SubscriptionsResponse(areaDtos);
     }
+
+    public SubscribingRegionTours mapToSubscribingRegionTours(Slice<Tour> tours) {
+
+        List<TourBasicDto> dtoList = tours.getContent().stream()
+                .map(t -> new TourBasicDto(
+                        t.getId(),
+                        t.getTitle(),
+                        t.getEventStartDate(),
+                        t.getEventEndDate(),
+                        areaCodeRegistry.getAreaName(t.getRegionCode().getAreaCode()),
+                        areaCodeRegistry.getSigunGuName(t.getRegionCode().getAreaCode(),t.getRegionCode().getSigunGuCode()),
+                        t.getMainImageUrl(),
+                        List.copyOf(t.getAllKeywords())
+                ))
+                .collect(Collectors.toList());
+
+        return new SubscribingRegionTours(dtoList.size(), tours.hasNext(), dtoList);
+    }
+
 
 
 }
